[addons] give child add-on class ptr to IAddonInstanceHandler
Seen during tests that the parent class pointer differ from child.
So that e.g. CImageDecoder is different to IAddonInstanceHandler. On tests has it before worked by callbacks and think the code makes something to get right place, but seems some additional work.
This gives now as constructor value about IAddonInstanceHandler the class pointer of child and to give to add-on about callbacks.