[addons] rename IAddonInstanceHandler to IInstanceHandler
Since it's already in a namespace called "ADDON" it wouldn't be necessary in classname anymore, in addition it shortens the name which can make "clang" happier.

Another background is at sandbox where these files are moved to different folders and namespaces.
There it would come as:
- xbmc/addons/interface/InstanceHandler.cpp
- xbmc/addons/interface/InstanceHandler.h
- KODI::ADDONS::INTERFACE::IInstanceHandler
Which again makes it longer (this makes clang go a little crazy if like before).